newscaster/api/requirements.txt
fastapi
uvicorn
python-dotenv
redis
pika
datetime
jinja2
prometheus_client
psycopg2-binary
migrate
requests